refactor(view): hooksの名前を修正
